Changements finaux classes

+++ Code/src/QuantiteElement.java
@@ -1,8 +1,8 @@
 
 public abstract class QuantiteElement {
 	private Element e;
-	private int quantite;
-	public QuantiteElement(Element e, int quantite) {
+	private double quantite;
+	public QuantiteElement(Element e, double quantite) {
 		this.e=e;
 		this.quantite=quantite;
 	}
