+++ Code/src/QuantitePlaqueAlu.java
@@ -1,7 +1,7 @@
 
 public class QuantitePlaqueAlu extends QuantiteElement{
 	
-	public QuantitePlaqueAlu(int quantite) {
+	public QuantitePlaqueAlu(double quantite) {
 		super(new PlaqueAlu(),quantite);
 	}
